test(nvsnap): fold path-safety tests into pathsafe_test.go
CI's BUILD-file check (#491) failed: pathsafe_write_test.go was a new file
absent from internal/agent/BUILD.bazel srcs.

Adding it to BUILD would work, but the package already pairs one test file per
source file and pathsafe_test.go was the obvious home. Same tests, no BUILD
churn, and one fewer place to look for coverage of pathsafe.go.
